getting violations from nys facilities
