+++ data/pagescrape.rb
@@ -1,19 +1,4 @@
-def cleanup(data)
-	def filter(datum)
-		return String(datum).gsub(/[^\w\s\-\/]/, "").gsub(/\r?\n|\r|\t/,"").gsub(/\s+/, " ").gsub(/<[^>]*>/, "")
-	end
-	data.each do |key, value|
-		if value.class == Hash
-			data[key] = cleanup(value)
-		elsif value.class == Array
-			value.each do |val|
-				cleanup(val)
-			end
-		else
-			data[key] = filter(value)
-		end
-	end
-end
+require_relative 'cleanup.rb'
 
 def pagescrape(page)
